requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
requests==2.31.0
pydantic==2.5.3
python-dotenv==1.0.0
claude-agent-sdk
